Fixed some float/double/int conversion warnings under Windows
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@1948 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 098848ba9e75b7b5c0d85c8f7b49774f642cb957
Former-commit-id: 84222bd7534b6d693e046a76b5647b8be25edce4

+++ src/drivers/usr_36GP/driver.cpp
@@ -872,7 +872,7 @@ void Driver::drive(tSituation *s)
 				car->_steerCmd = cmd_steer;
 				car->_gearCmd = cmd_gear;
 				car->_clutchCmd = cmd_clutch;
-				car->_lightCmd = cmd_light;
+				car->_lightCmd = (int)cmd_light;
 				return;
 			}
 		}
@@ -942,7 +942,7 @@ void Driver::drive(tSituation *s)
 	cmd_steer = car->_steerCmd;
 	cmd_clutch = car->_clutchCmd;
 	cmd_gear = car->_gearCmd;
-	cmd_light = car->_lightCmd;
+	cmd_light = (float)car->_lightCmd;
 }
